Fix force sync of large vaults from Obsidian
Previously if you tried to force sync a vault with more than 1000
files it would only end up keeping the last batch because the PUT API
call would delete all previous entries.

This change ensures the PUT call is only made for the first batch of a
force sync (regenerate) request. The rest of the API calls are PATCH,
so they do not delete files from the previous batches